Fixes #82.
Eliminate the outlet for annotateButton, as the only reason we needed it was to set the image inside.  Set the image statically in the NIB instead.  This removes the annotate + caution icon.  A separate icon may be unnecessary since the popover is directly attached to the modality featuring a badge, added in 841f49ae1e5916f3a4f8000c66cf6b3b2907f08c.