Fixes #56.
The `editing` property of the cell gets set to NO by the GMGridView after you delete a sensor, and since that is the superclass, the subclass WSPersonTableViewCell's mutator which sets the style of the `editButton` doesn't get called.  Fortunately, the `editing` property doesn't matter for the operation (more of a formality), so ignore it and set the style based on the `selected` property of `editButton`.