Move setup of notification listeners from viewDidLoad to viewWillAppear:.
viewDidLoad is only called once for a view loaded into a PopoverController, so the CaptureController notification listeners were never instantiated after the first time the CaptureController was loaded.  This causes behavior where a downloaded image will post as a thumbnail in the person list but not in the CaptureController (maybe #117?) and a cleared image could not be updated until closing (#132).