Implements #80 -- "lightbox mode."
Double-tap a capture to view it full-screen.  Single-tap on the outside of the image to dismiss.

Known-issue: Device rotation messages are ignored during lightboxing.  Rotations cause popovers to close, and since the CaptureController popover is the presenting ViewController for the lightbox, the lightbox will also close.  It's too visually jarring to restore the popover and lightbox in my opinion.