Make the sessionID read-only.
This brings up an interesting point: WSDeviceLink seems to be an object that just handles network communications, and so it takes a sessionID and sourceObjectID with every API call.  However, we store the session ID so that we can chain together commands, which breaks this idea.  This needs to change (related to github issue #148).